Add ellipsizeMode parameter to wxRenderNative::DrawItemText().
Instead of the default end ellipsize mode used in the native and generic implementation, allow specifying the mode with an additional parameter. wxEllipsizeMode has been moved from controls.h to gdicmn.h.
