+++ src/generic/renderg.cpp
@@ -150,7 +150,8 @@ public:
                               const wxString& text,
                               const wxRect& rect,
                               int align = wxALIGN_LEFT | wxALIGN_TOP,
-                              int flags = 0) wxOVERRIDE;
+                              int flags = 0,
+                              wxEllipsizeMode ellipsizeMode = wxELLIPSIZE_END) wxOVERRIDE;
 
     virtual wxSplitterRenderParams GetSplitterParams(const wxWindow *win) wxOVERRIDE;
 
@@ -888,7 +889,8 @@ wxRendererGeneric::DrawItemText(wxWindow* win,
                                 const wxString& text,
                                 const wxRect& rect,
                                 int align,
-                                int flags)
+                                int flags,
+                                wxEllipsizeMode ellipsizeMode)
 {
     // Determine text color
     wxColour textColour;
@@ -913,7 +915,7 @@ wxRendererGeneric::DrawItemText(wxWindow* win,
     }
 
     const wxString paintText = wxControl::Ellipsize(text, dc,
-                                                    wxELLIPSIZE_END,
+                                                    ellipsizeMode,
                                                     rect.GetWidth());
 
     // Draw text
